Added GLS and CGMS parsers.

+++ lib_service/src/main/java/no/nordicsemi/android/service/services/CGMManager.kt
@@ -222,7 +222,7 @@ internal class CGMManager : ServiceManager {
                     .write(
                         if (state.value.records.isNotEmpty()) {
                             RecordAccessControlPointInputParser.reportStoredRecordsGreaterThenOrEqualTo(
-                                highestSequenceNumber
+                                highestSequenceNumber.toShort()
                             )
                         } else {
                             RecordAccessControlPointInputParser.reportAllStoredRecords()
